requirements.txt
mcp>=0.1.0
uvicorn>=0.20.0
fastapi>=0.90.0
supabase>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
